+++ bin/miden-cli/src/commands/init.rs
@@ -27,7 +27,7 @@ const BASIC_AUTH_TEMPLATE_FILE: &[u8] =
 // ================================================================================================
 
 #[derive(Debug, Clone, Parser)]
-#[clap(
+#[command(
     about = "Initialize the client. It will create a file named `miden-client.toml` that holds \
 the CLI and client configurations, and will be placed by default in the current working \
 directory"
@@ -39,14 +39,14 @@ pub struct InitCmd {
     network: Network,
 
     /// Path to the store file.
-    #[clap(long)]
+    #[arg(long)]
     store_path: Option<String>,
 
     /// RPC endpoint for the proving service. Required if proving mode is set to remote.
     /// The endpoint must be in the form of "{protocol}://{hostname}:{port}", being the protocol
     /// and port optional.
     /// If the proving RPC isn't set, the proving mode will be set to local.
-    #[clap(long)]
+    #[arg(long)]
     remote_prover_endpoint: Option<String>,
 
     /// Maximum number of blocks the client can be behind the network.