+++ bin/miden-cli/src/lib.rs
@@ -49,19 +49,19 @@ const TX_GRACEFUL_BLOCK_DELTA: u32 = 20;
 
 /// Root CLI struct.
 #[derive(Parser, Debug)]
-#[clap(
+#[command(
     name = "miden-client",
     about = "The Miden client",
     version,
     rename_all = "kebab-case"
 )]
 pub struct Cli {
-    #[clap(subcommand)]
+    #[command(subcommand)]
     action: Command,
 
     /// Activates the executor's debug mode, which enables debug output for scripts
     /// that were compiled and executed with this mode.
-    #[clap(short, long, default_value_t = false)]
+    #[arg(short, long, default_value_t = false)]
     debug: bool,
 }
 
@@ -79,7 +79,7 @@ pub enum Command {
     /// View a summary of the current client state.
     Info,
     Tags(TagsCmd),
-    #[clap(name = "tx")]
+    #[command(name = "tx")]
     Transaction(TransactionCmd),
     Mint(MintCmd),
     Send(SendCmd),