+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v0.9.4 (2025-07-01)
+
+* Support Operations From Counter Contract FPI Example in Web Client (#958).
+
 ## v0.9.3 (2025-06-28)
 
 * Fixed a bug where some partial MMR nodes were missing and causing problems with note consumption (#995).